Updated manifest
