+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
             android:name="com.facebook.sdk.ApplicationId"
             android:value="@string/app_id" />
         <activity
-            android:name=".MainActivity"
+            android:name=".Register"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
